+++ api/Core.Model/DTO/AirGradientMeasureDto.cs
@@ -3,7 +3,7 @@ namespace Core.Model.DTO;
 public class AirGradientMeasureDto
 {
     public int Wifi { get; set; }
-    public int Rco2 { get; set; }
+    public decimal Rco2 { get; set; }
     public decimal Pm01 { get; set; }
     public decimal Pm02 { get; set; }
     public decimal? Pm02Compensated { get; set; }
@@ -12,6 +12,6 @@ public class AirGradientMeasureDto
     public decimal? AtmpCompensated { get; set; }
     public decimal Rhum { get; set; }
     public decimal? RhumCompensated { get; set; }
-    public int TvocIndex { get; set; }
-    public int NoxIndex { get; set; }
+    public decimal TvocIndex { get; set; }
+    public decimal NoxIndex { get; set; }
 }
